Create README.md
Repository containing code for Dissertation for MSc Social and Geographic DataScience for candidate ZJXV1.
NB: we cannot provide the full data used as it is property of Zoopla. If data is necessary, this can be request on behalf of the institution (UCL). Furthermore, this is also applicable for certain code, as it can not directly be shared without confirmation from HR